+++ package.json
@@ -2,7 +2,9 @@
   "dependencies": {
     "check-node-version": "^4.2.1",
     "crypto-js": "^4.2.0",
-    "jszip": "^3.10.1"
+    "fs": "^0.0.1-security",
+    "jszip": "^3.10.1",
+    "path": "^0.12.7"
   },
   "devDependencies": {
     "js-cookie": "^3.0.5"
